move some utils to pkg manager/utils
