+++ pkg/manager/member/dm_worker_member_manager_test.go
@@ -20,6 +20,7 @@ import (
 	"testing"
 
 	"github.com/pingcap/tidb-operator/pkg/apis/label"
+	mngerutils "github.com/pingcap/tidb-operator/pkg/manager/utils"
 
 	"github.com/google/go-cmp/cmp"
 	"k8s.io/apimachinery/pkg/util/intstr"
@@ -654,7 +655,7 @@ func TestWorkerSyncConfigUpdate(t *testing.T) {
 				g.Expect(r.listCm).To(Succeed())
 				g.Expect(r.cms).To(HaveLen(2))
 				g.Expect(r.getSet).To(Succeed())
-				using := FindConfigMapVolume(&r.set.Spec.Template.Spec, func(name string) bool {
+				using := mngerutils.FindConfigMapVolume(&r.set.Spec.Template.Spec, func(name string) bool {
 					return strings.HasPrefix(name, controller.DMWorkerMemberName("test"))
 				})
 				g.Expect(using).NotTo(BeEmpty())
