+++ pkg/manager/member/ticdc_upgrader_test.go
@@ -16,10 +16,13 @@ package member
 import (
 	"testing"
 
-	. "github.com/onsi/gomega"
+	mngerutils "github.com/pingcap/tidb-operator/pkg/manager/utils"
+
 	"github.com/pingcap/tidb-operator/pkg/apis/label"
 	"github.com/pingcap/tidb-operator/pkg/apis/pingcap/v1alpha1"
 	"github.com/pingcap/tidb-operator/pkg/controller"
+
+	. "github.com/onsi/gomega"
 	apps "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -64,7 +67,7 @@ func TestTiCDCUpgrader_Upgrade(t *testing.T) {
 		if test.changeOldSet != nil {
 			test.changeOldSet(oldSet)
 		}
-		SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+		mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 
 		err := upgrader.Upgrade(tc, oldSet, newSet)
 		if test.errorExpect {
@@ -179,7 +182,7 @@ func TestTiCDCUpgrader_Upgrade(t *testing.T) {
 				tc.Status.TiCDC.Synced = true
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) {
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 			},
 			errorExpect: false,
 			expectFn: func(g *GomegaWithT, tc *v1alpha1.TidbCluster, newSet *apps.StatefulSet) {
@@ -198,7 +201,7 @@ func TestTiCDCUpgrader_Upgrade(t *testing.T) {
 				tc.Status.TiCDC.Synced = true
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) {
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 			},
 			errorExpect: false,
 			expectFn: func(g *GomegaWithT, tc *v1alpha1.TidbCluster, newSet *apps.StatefulSet) {
@@ -217,7 +220,7 @@ func TestTiCDCUpgrader_Upgrade(t *testing.T) {
 				tc.Status.TiCDC.Synced = true
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) {
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 			},
 			errorExpect: false,
 			expectFn: func(g *GomegaWithT, tc *v1alpha1.TidbCluster, newSet *apps.StatefulSet) {
@@ -236,7 +239,7 @@ func TestTiCDCUpgrader_Upgrade(t *testing.T) {
 				tc.Status.TiCDC.Synced = true
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) {
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 			},
 			errorExpect: false,
 			expectFn: func(g *GomegaWithT, tc *v1alpha1.TidbCluster, newSet *apps.StatefulSet) {
@@ -255,7 +258,7 @@ func TestTiCDCUpgrader_Upgrade(t *testing.T) {
 				tc.Status.TiCDC.Synced = true
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) {
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 			},
 			errorExpect: false,
 			expectFn: func(g *GomegaWithT, tc *v1alpha1.TidbCluster, newSet *apps.StatefulSet) {
