+++ pkg/manager/member/tidb_upgrader_test.go
@@ -16,10 +16,13 @@ package member
 import (
 	"testing"
 
-	. "github.com/onsi/gomega"
+	mngerutils "github.com/pingcap/tidb-operator/pkg/manager/utils"
+
 	"github.com/pingcap/tidb-operator/pkg/apis/label"
 	"github.com/pingcap/tidb-operator/pkg/apis/pingcap/v1alpha1"
 	"github.com/pingcap/tidb-operator/pkg/controller"
+
+	. "github.com/onsi/gomega"
 	apps "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -61,7 +64,7 @@ func TestTiDBUpgrader_Upgrade(t *testing.T) {
 		if test.getLastAppliedConfigErr {
 			oldSet.SetAnnotations(map[string]string{LastAppliedConfigAnnotation: "fake apply config"})
 		} else {
-			SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+			mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 		}
 		err := upgrader.Upgrade(tc, oldSet, newSet)
 		if test.errorExpect {
