+++ pkg/manager/member/tiflash_upgrader_test.go
@@ -26,6 +26,8 @@ import (
 	podinformers "k8s.io/client-go/informers/core/v1"
 	"k8s.io/utils/pointer"
 
+	mngerutils "github.com/pingcap/tidb-operator/pkg/manager/utils"
+
 	"github.com/pingcap/tidb-operator/pkg/apis/label"
 	"github.com/pingcap/tidb-operator/pkg/apis/pingcap/v1alpha1"
 	"github.com/pingcap/tidb-operator/pkg/controller"
@@ -91,7 +93,7 @@ func TestTiFlashUpgraderUpgrade(t *testing.T) {
 			name:     "modify oldSet update strategy to OnDelete",
 			changeFn: nil,
 			changeOldSet: func(oldSet *apps.StatefulSet) {
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 				oldSet.Spec.UpdateStrategy = apps.StatefulSetUpdateStrategy{
 					Type: apps.OnDeleteStatefulSetStrategyType,
 				}
@@ -110,7 +112,7 @@ func TestTiFlashUpgraderUpgrade(t *testing.T) {
 			name:     "set oldSet's RollingUpdate strategy to nil",
 			changeFn: nil,
 			changeOldSet: func(oldSet *apps.StatefulSet) {
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 				oldSet.Spec.UpdateStrategy = apps.StatefulSetUpdateStrategy{
 					Type: apps.RollingUpdateStatefulSetStrategyType,
 				}
@@ -137,7 +139,7 @@ func TestTiFlashUpgraderUpgrade(t *testing.T) {
 				tc.Status.TiFlash.Stores["3"] = store
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) {
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 			},
 			changePods:   nil,
 			updatePodErr: false,
@@ -163,7 +165,7 @@ func TestTiFlashUpgraderUpgrade(t *testing.T) {
 				tc.Status.TiFlash.Stores["2"] = store
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) {
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 				oldSet.Status.CurrentReplicas = 2
 				oldSet.Status.UpdatedReplicas = 1
 				oldSet.Spec.UpdateStrategy.RollingUpdate.Partition = pointer.Int32Ptr(2)
@@ -186,7 +188,7 @@ func TestTiFlashUpgraderUpgrade(t *testing.T) {
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) {
 				oldSet.Spec.Template.Spec.Containers[0].Image = "old-image"
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 			},
 			changePods:   nil,
 			updatePodErr: false,
@@ -224,7 +226,7 @@ func TestTiFlashUpgraderUpgrade(t *testing.T) {
 				tc.Status.TiFlash.Synced = true
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) {
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 			},
 			changePods:   nil,
 			updatePodErr: false,
@@ -272,7 +274,7 @@ func TestTiFlashUpgraderUpgrade(t *testing.T) {
 				})
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) { // tigger upgrade
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 				oldSet.Spec.Template.Spec.Containers[0].Image = "old-image"
 			},
 			changePods: func(pods []*corev1.Pod, tc *v1alpha1.TidbCluster, old, new *apps.StatefulSet) {
@@ -303,7 +305,7 @@ func TestTiFlashUpgraderUpgrade(t *testing.T) {
 				})
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) { // tigger upgrade
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 				oldSet.Spec.Template.Spec.Containers[0].Image = "old-image"
 			},
 			changePods: func(pods []*corev1.Pod, tc *v1alpha1.TidbCluster, old, new *apps.StatefulSet) {
@@ -334,7 +336,7 @@ func TestTiFlashUpgraderUpgrade(t *testing.T) {
 				})
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) { // tigger upgrade
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 				oldSet.Spec.Template.Spec.Containers[0].Image = "old-image"
 			},
 			changePods: func(pods []*corev1.Pod, tc *v1alpha1.TidbCluster, old, new *apps.StatefulSet) {
@@ -365,7 +367,7 @@ func TestTiFlashUpgraderUpgrade(t *testing.T) {
 				})
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) { // tigger upgrade
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 				oldSet.Spec.Template.Spec.Containers[0].Image = "old-image"
 			},
 			changePods: func(pods []*corev1.Pod, tc *v1alpha1.TidbCluster, old, new *apps.StatefulSet) {
@@ -396,7 +398,7 @@ func TestTiFlashUpgraderUpgrade(t *testing.T) {
 				})
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) { // tigger upgrade
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 				oldSet.Spec.Template.Spec.Containers[0].Image = "old-image"
 			},
 			changePods: func(pods []*corev1.Pod, tc *v1alpha1.TidbCluster, old, new *apps.StatefulSet) {
@@ -427,7 +429,7 @@ func TestTiFlashUpgraderUpgrade(t *testing.T) {
 				})
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) { // tigger upgrade
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 				oldSet.Spec.Template.Spec.Containers[0].Image = "old-image"
 			},
 			changePods: func(pods []*corev1.Pod, tc *v1alpha1.TidbCluster, old, new *apps.StatefulSet) {
@@ -459,7 +461,7 @@ func TestTiFlashUpgraderUpgrade(t *testing.T) {
 				})
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) { // tigger upgrade
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 				oldSet.Spec.Template.Spec.Containers[0].Image = "old-image"
 			},
 			changePods: func(pods []*corev1.Pod, tc *v1alpha1.TidbCluster, old, new *apps.StatefulSet) {
@@ -491,7 +493,7 @@ func TestTiFlashUpgraderUpgrade(t *testing.T) {
 				})
 			},
 			changeOldSet: func(oldSet *apps.StatefulSet) { // tigger upgrade
-				SetStatefulSetLastAppliedConfigAnnotation(oldSet)
+				mngerutils.SetStatefulSetLastAppliedConfigAnnotation(oldSet)
 				oldSet.Spec.Template.Spec.Containers[0].Image = "old-image"
 			},
 			changePods: func(pods []*corev1.Pod, tc *v1alpha1.TidbCluster, old, new *apps.StatefulSet) {
