+++ pkg/manager/utils/configmap.go
@@ -1,0 +1,51 @@
+// Copyright 2020 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package utils
+
+import (
+	"crypto/sha256"
+	"encoding/json"
+	"fmt"
+
+	corev1 "k8s.io/api/core/v1"
+)
+
+func AddConfigMapDigestSuffix(cm *corev1.ConfigMap) error {
+	sum, err := Sha256Sum(cm.Data)
+	if err != nil {
+		return err
+	}
+	suffix := fmt.Sprintf("%x", sum)[0:7]
+	cm.Name = fmt.Sprintf("%s-%s", cm.Name, suffix)
+	return nil
+}
+
+func Sha256Sum(v interface{}) (string, error) {
+	data, err := json.Marshal(v)
+	if err != nil {
+		return "", err
+	}
+	sum := sha256.Sum256(data)
+	return fmt.Sprintf("%x", sum), nil
+}
+
+// FindConfigMapVolume returns the configmap which's name matches the predicate in a PodSpec, empty indicates not found
+func FindConfigMapVolume(podSpec *corev1.PodSpec, pred func(string) bool) string {
+	for _, vol := range podSpec.Volumes {
+		if vol.ConfigMap != nil && pred(vol.ConfigMap.LocalObjectReference.Name) {
+			return vol.ConfigMap.LocalObjectReference.Name
+		}
+	}
+	return ""
+}
