+++ pkg/manager/utils/statefulset_builder.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package statefulset
+package utils
 
 import (
 	"github.com/pingcap/tidb-operator/pkg/util"
