+++ pkg/monitor/monitor/monitor_manager.go
@@ -29,6 +29,7 @@ import (
 	"github.com/pingcap/tidb-operator/pkg/monitor"
 	"github.com/pingcap/tidb-operator/pkg/pdapi"
 	"github.com/pingcap/tidb-operator/pkg/util"
+	mngerutils "github.com/pingcap/tidb-operator/pkg/manager/utils"
 	utildiscovery "github.com/pingcap/tidb-operator/pkg/util/discovery"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -248,7 +249,7 @@ func (m *MonitorManager) syncTidbMonitorStatefulset(tc *v1alpha1.TidbCluster, dc
 		}
 		setNotExist := errors.IsNotFound(err)
 		if setNotExist {
-			err = member.SetStatefulSetLastAppliedConfigAnnotation(newMonitorSts)
+			err = mngerutils.SetStatefulSetLastAppliedConfigAnnotation(newMonitorSts)
 			if err != nil {
 				return err
 			}
@@ -258,7 +259,7 @@ func (m *MonitorManager) syncTidbMonitorStatefulset(tc *v1alpha1.TidbCluster, dc
 			isAllCreated = false
 			continue
 		}
-		err = member.UpdateStatefulSet(m.deps.StatefulSetControl, monitor, newMonitorSts, oldMonitorSetTmp)
+		err = mngerutils.UpdateStatefulSet(m.deps.StatefulSetControl, monitor, newMonitorSts, oldMonitorSetTmp)
 		if err != nil {
 			klog.Errorf("Fail to update statefulset[%s/%s] for tm [%s/%s], err: %v", ns, stsName, ns, name, err)
 			return err
